Move ProcessorRule into separate module: processor-testlib
New module allows to test ValueType / ValueAttribute directly from unit test
