+++ pom.xml
@@ -45,6 +45,7 @@
     <module>mongo</module>
     <module>func</module>
     <module>value-processor</module>
+    <module>processor-testlib</module>
     <module>value-annotations</module>
     <module>value-fixture</module>
     <module>value</module>
@@ -111,6 +112,7 @@
     <mongo-java-server.version>1.14.0</mongo-java-server.version>
     <jackson.version>2.8.11</jackson.version>
     <jackson-databind.version>2.8.11.3</jackson-databind.version>
+    <junit.version>4.12</junit.version>
     <mockito.verion>1.10.19</mockito.verion>
     <guava.version>19.0</guava.version>
     <gson.version>2.8.5</gson.version>
