+++ processor-testlib/README.md
@@ -1,0 +1,4 @@
+Test of utilities to test immutable processor (apt).
+Specifically classes like `ValueType` or `ValueAttribute`.
+ 
+
