+++ processor-testlib/test/org/immutables/value/processor/meta/ProcessorRuleTest.java
