+++ value-processor/pom.xml
@@ -77,12 +77,6 @@
       <version>${project.version}</version>
       <scope>test</scope>
     </dependency>
-    <dependency>
-      <groupId>com.google.testing.compile</groupId>
-      <artifactId>compile-testing</artifactId>
-      <version>0.18</version>
-      <scope>test</scope>
-    </dependency>
   </dependencies>
 
 </project>
